api: 내 정보 조회 api 연결 (#402)
