+++ src/pages/home/apis/axios.ts
@@ -7,12 +7,6 @@ export const getAdvertisements = async () => {
   return data;
 };
 
-export const getMyPage = async () => {
-  const { data } = await instance.get(API_URL.MEMBERS_ME);
-
-  return data;
-};
-
 export const postLogout = async () => {
   const { data } = await instance.post(API_URL.AUTH_LOGOUT);
