+++ src/pages/mypage/MyPage.tsx
@@ -1,5 +1,5 @@
-// import { useGetMyPage } from '@/pages/home/apis/queries';
-// import { DEFAULT_USER_DATA } from '@/pages/home/constants';
+import { DEFAULT_USER_DATA } from '@/pages/home/constants';
+import { useGetMyPage } from '@/pages/mypage/apis/queries';
 import TabWrapper from '@/pages/mypage/components/TabWrapper/TabWrapper';
 import * as styles from '@/pages/mypage/myPage.css';
 import { getUserRole, setUserRole } from '@/pages/mypage/utils/storage';
@@ -22,7 +22,8 @@ const MyPage = () => {
 
   // const isInstructor = roleData === 'TEACHER';
 
-  //  const { data: userData = DEFAULT_USER_DATA } = useGetMyPage({});
+  const { data: userData = DEFAULT_USER_DATA } = useGetMyPage({});
+  console.log(userData);
 
   return (
     <main className={styles.containerStyle}>
