+++ src/pages/mypage/apis/axios.ts
@@ -1,0 +1,8 @@
+import { instance } from '@/shared/apis/instance';
+import { API_URL } from '@/shared/constants/apiURL';
+
+export const getMyPage = async () => {
+  const { data } = await instance.get(API_URL.MEMBERS_ME);
+
+  return data;
+};
