+++ src/pages/mypage/components/TabWrapper/components/StudentContent/StudentContent.tsx
@@ -1,3 +1,4 @@
+import { useGetMyPage } from '@/pages/mypage/apis/queries';
 import ClassProgress from '@/pages/mypage/components/TabWrapper/components/StudentContent/components/ClassProgress/ClassProgress';
 import MenuButton from '@/pages/mypage/components/TabWrapper/components/StudentContent/components/MenuButton/MenuButton';
 import * as styles from '@/pages/mypage/components/TabWrapper/components/StudentContent/studentContent.css';
@@ -7,10 +8,13 @@ import Divider from '@/shared/components/Divider/Divider';
 import InfoComponent from '@/shared/components/InfoComponent/InfoComponent';
 import Text from '@/shared/components/Text/Text';
 import BottomList from '../../../BottomList/BottomList';
-import { mockMyPageData } from '../../mockData';
 
 const StudentContent = () => {
-  const data = mockMyPageData;
+  const { data, isLoading } = useGetMyPage({});
+
+  if (isLoading || !data) {
+    return <></>;
+  }
 
   return (
     <div className={styles.containerStyle}>
