+++ src/pages/mypage/types/api.ts
@@ -3,3 +3,10 @@ export interface LessonCountResponseTypes {
   duringLessonCount: number;
   afterLessonCount: number;
 }
+
+export interface MyPageResponseTypes {
+  profileImageUrl: string;
+  nickname: string;
+  name: string;
+  phoneNumber: string;
+}
