+++ src/shared/types/myPageTypes.ts
@@ -1,10 +1,3 @@
 export interface RoleNameResponseTypes {
   role: string;
 }
-export interface MyPageResponseTypes {
-  profileImageUrl: string;
-  nickname: string;
-  reservationCount: number;
-  favoriteCount: number;
-  lessonCount: number | null;
-}
